Implemented Databinding

+++ QuickStop.Components/Helpers/HotelRoomBooking.cs
@@ -6,7 +6,7 @@ namespace QuickStop.Components.Helpers
     {
         public static decimal CalculateTotalPrice(decimal hotelPrice, int guestCount, int minGuestCount, int maxGuestCount, DateTime checkIn, DateTime checkOut)
         {
-            return hotelPrice * (1 + ((guestCount - minGuestCount) / (maxGuestCount - minGuestCount))) * (checkOut - checkIn).Days;
+            return hotelPrice * (1 + ((decimal)(guestCount - minGuestCount) / (maxGuestCount - minGuestCount))) * (checkOut - checkIn).Days;
         }
     }
 }
